+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ## [2.6.0] - 2018-09-13
 ### Removed
+- cmds_gen.Is_Add_Cmd:  Function was previously deprecated.
 - cmds_gen.Disconnect:  Function was previously deprecated.
 - arg_parser.Arg_Set_Path:  Function was previously deprecated.